requirements.txt
python-dotenv
pip install -r requirements.txt